ADD code refactoring on producers and consumers with usage() and testing crashing Kafka brokers to test fault tolerance

+++ src/bash/consumer_listen_topic.sh
@@ -3,6 +3,18 @@
 ###
 # Must provide the broker ID as $1, the broker port as $2 and the topic name as $3
 
+function usage() {
+    echo "usage: ${0} <broker ID> <broker port> <topic name>"
+}
+
+if [ -z "$1" -o -z "$2" -o -z "$3" ]; then
+    usage
+    exit 1
+fi
+
 docker run --rm ches/kafka \
-    kafka-console-consumer.sh --bootstrap-server $(docker inspect kafka$1 -f '{{.NetworkSettings.IPAddress}}'):$2 \
-    --from-beginning --topic $3
+    kafka-console-consumer.sh --new-consumer \
+    --bootstrap-server $(docker inspect kafka$1 -f '{{.NetworkSettings.IPAddress}}'):$2 \
+    --topic $3
+
+#    --from-beginning \
